refactor: move implementation helpers out of averagingAnalyser.types.ts (M1)
Move createAccumulator, createDataPointAccumulator, and accumToMetric from
.types.ts to .accumulation.ts. .types.ts now exports only types. Update
imports in averagingAnalyser.ts and averagingAnalyser.rows.ts.
