+++ src/frontend/src/services/dataAnalysis/analysers/averagingAnalyser.accumulation.ts
@@ -1,7 +1,51 @@
-import type { AveragingAnalyserInput } from '../dataAnalysis.zod';
+import type { AveragingAnalyserInput, MetricResult } from '../dataAnalysis.zod';
 import type { CriterionWeightings } from './averagingAnalyser';
-import type { AssessmentScore, DataPointAccumulator } from './averagingAnalyser.types';
-import { createDataPointAccumulator } from './averagingAnalyser.types';
+import type {
+  AssessmentScore,
+  DataPointAccumulator,
+  MetricAccumulator,
+} from './averagingAnalyser.types';
+
+/**
+ * Create a zeroed-out metric accumulator.
+ *
+ * @returns {MetricAccumulator} A fresh entry with all fields at zero.
+ */
+export function createAccumulator(): MetricAccumulator {
+  return { weightedSum: 0, totalWeight: 0, applicableDataPoints: 0, totalDataPoints: 0 };
+}
+
+/**
+ * Create a zeroed-out data-point accumulator set.
+ *
+ * @returns {DataPointAccumulator} A fresh entry with all sub-accumulators at zero.
+ */
+export function createDataPointAccumulator(): DataPointAccumulator {
+  return {
+    completeness: createAccumulator(),
+    accuracy: createAccumulator(),
+    spag: createAccumulator(),
+    overall: createAccumulator(),
+  };
+}
+
+/**
+ * Convert a metric accumulator to its output MetricResult shape.
+ *
+ * @param {MetricAccumulator} accumulator - The mutable accumulator to convert.
+ * @returns {MetricResult} A read-only MetricResult snapshot.
+ */
+export function accumToMetric(accumulator: MetricAccumulator): MetricResult {
+  return {
+    value:
+      accumulator.applicableDataPoints === 0
+        ? null
+        : accumulator.weightedSum / accumulator.totalWeight,
+    totalWeight: accumulator.totalWeight,
+    applicableDataPoints: accumulator.applicableDataPoints,
+    totalDataPoints: accumulator.totalDataPoints,
+  };
+}
 
 /**
  * Pre-register tasks so entries with zero submissions appear in perTask.
